+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 - Added on_mouse_button callback (#184, #185)
 - Added on_tick callback (#187)
 - Added on_mouse_bounds callback (#181, #182)
+- Added inject_event_handler macro to handle default events for callbacks (#192)
 
 ## 0.4.5